Changed the background of the main swipe screen to swipe_screen_lowres.png. Please note that the high res version of this image is causing severe overloading of the main process thread, so don't use this image as the background until we can properly multi-thread the swipe activity.
Added debug.keystore to app\external. This file is needed in order to use the Google Maps API key that is set up for this project.

Changed the background colors of the Driver and Passenger profile pages to match the new color scheme.

Added a map button to the action bar which will launch MapActivity. Note that you must have Google Play Services installed on your emulator device for Google Maps to load properly.